Added numpy operations module
